huge improvements

+++ lib/worker_killer/killer/delayed_job.rb
@@ -2,23 +2,19 @@ module WorkerKiller
   module Killer
     class DelayedJob < ::WorkerKiller::Killer::Base
 
-      def do_kill(sig, pid, alive_sec, dj:, **_params)
-        if sig == :KILL
-          logger.error "#{self.class}: force to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (trial #{kill_attempts})"
+      def do_kill(sig, pid, alive_sec, dj:, **_kwargs)
+        case sig
+        when :QUIT
+          logger.info "#{self.class}: try to stop DelayedJob due to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (kill attempts #{kill_attempts})"
+          dj.stop
+        when :TERM
+          logger.warn "#{self.class}: force to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (kill attempts #{kill_attempts})"
           Process.kill sig, pid
-          return
-        end
-
-        dj.stop
-        logger.info "#{self.class}: try to stop DelayedJob due to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (trial #{kill_attempts})"
-
-        return if sig != :TERM
-
-        if @termination
-          logger.warn "#{self.class}: force to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (trial #{kill_attempts})"
+        when :KILL
+          logger.error "#{self.class}: force to #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (kill attempts #{kill_attempts})"
           Process.kill sig, pid
         else
-          @termination = true
+          logger.error "#{self.class}: DO NOTHING: unknown #{sig} self (pid: #{pid}) alive: #{alive_sec} sec (kill attempts #{kill_attempts})"
         end
       end
